// questionnaire/src/protos/gen/questionnaire/UpdateCategory.ts
// Original file: src/protos/category.proto


export interface UpdateCategory {
  'access_token'?: (string);
  'name'?: (string);
  'status'?: (string);
  'id'?: (string);
}

export interface UpdateCategory__Output {
  'access_token': (string);
  'name': (string);
  'status': (string);
  'id': (string);
}